+++ src/generators/identifier-names-generators/AbstractIdentifierNamesGenerator.ts
@@ -1,6 +1,8 @@
 import { inject, injectable } from 'inversify';
 import { ServiceIdentifiers } from '../../container/ServiceIdentifiers';
 
+import { TNodeWithLexicalScope } from '../../types/node/TNodeWithLexicalScope';
+
 import { IIdentifierNamesGenerator } from '../../interfaces/generators/identifier-names-generators/IIdentifierNamesGenerator';
 import { IOptions } from '../../interfaces/options/IOptions';
 import { IRandomGenerator } from '../../interfaces/utils/IRandomGenerator';
@@ -23,6 +25,11 @@ export abstract class AbstractIdentifierNamesGenerator implements IIdentifierNam
     protected readonly preservedNamesSet: Set<string> = new Set();
 
     /**
+     * @type {Map<TNodeWithLexicalScope, Set<string>>}
+     */
+    protected readonly lexicalScopesPreservedNamesMap: Map<TNodeWithLexicalScope, Set<string>> = new Map();
+
+    /**
      * @param {IRandomGenerator} randomGenerator
      * @param {IOptions} options
      */
@@ -36,10 +43,17 @@ export abstract class AbstractIdentifierNamesGenerator implements IIdentifierNam
 
     /**
      * @param {string} name
-     * @returns {void}
+     * @param {TNodeWithLexicalScope} lexicalScopeNode
      */
-    public preserveName (name: string): void {
+    public preserveName (name: string, lexicalScopeNode: TNodeWithLexicalScope): void {
         this.preservedNamesSet.add(name);
+
+        const preservedNamesForLexicalScopeSet: Set<string> =
+            this.lexicalScopesPreservedNamesMap.get(lexicalScopeNode) ?? new Set();
+
+        preservedNamesForLexicalScopeSet.add(name);
+
+        this.lexicalScopesPreservedNamesMap.set(lexicalScopeNode, preservedNamesForLexicalScopeSet);
     }
 
     /**
@@ -48,6 +62,26 @@ export abstract class AbstractIdentifierNamesGenerator implements IIdentifierNam
      */
     public isValidIdentifierName (name: string): boolean {
         return this.notReservedName(name) && !this.preservedNamesSet.has(name);
+    }
+
+    /**
+     * @param {string} name
+     * @param {TNodeWithLexicalScope} lexicalScopeNode
+     * @returns {boolean}
+     */
+    public isValidIdentifierNameInLexicalScope (name: string, lexicalScopeNode: TNodeWithLexicalScope): boolean {
+        if (!this.notReservedName(name)) {
+            return false;
+        }
+
+        const preservedNamesForLexicalScopeSet: Set<string> | null =
+            this.lexicalScopesPreservedNamesMap.get(lexicalScopeNode) ?? null;
+
+        if (!preservedNamesForLexicalScopeSet) {
+            return true;
+        }
+
+        return !preservedNamesForLexicalScopeSet.has(name);
     }
 
     /**
@@ -68,6 +102,13 @@ export abstract class AbstractIdentifierNamesGenerator implements IIdentifierNam
      * @returns {string}
      */
     public abstract generate (nameLength?: number): string;
+
+    /**
+     * @param {TNodeWithLexicalScope} lexicalScopeNode
+     * @param {number} nameLength
+     * @returns {string}
+     */
+    public abstract generateForLexicalScope (lexicalScopeNode: TNodeWithLexicalScope, nameLength?: number): string;
 
     /**
      * @param {number} nameLength