+++ src/interfaces/generators/identifier-names-generators/IIdentifierNamesGenerator.ts
@@ -1,9 +1,19 @@
+import { TNodeWithLexicalScope } from '../../../types/node/TNodeWithLexicalScope';
+
 export interface IIdentifierNamesGenerator {
     /**
      * @param {number} nameLength
      * @returns {string}
      */
     generate (nameLength?: number): string;
+
+    /**
+     * @param {TNodeWithLexicalScope} lexicalScopeNode
+     * @param {number} nameLength
+     * @returns {string}
+     */
+    generateForLexicalScope (lexicalScopeNode: TNodeWithLexicalScope, nameLength?: number): string;
+
 
     /**
      * @param {number} nameLength
@@ -18,7 +28,15 @@ export interface IIdentifierNamesGenerator {
     isValidIdentifierName (identifierName: string): boolean;
 
     /**
-     * @param {string} name
+     * @param {string} identifierName
+     * @param {TNodeWithLexicalScope} lexicalScopeNode
+     * @returns {boolean}
      */
-    preserveName (name: string): void;
+    isValidIdentifierNameInLexicalScope (identifierName: string, lexicalScopeNode: TNodeWithLexicalScope): boolean;
+
+    /**
+     * @param {string} identifierName
+     * @param {TNodeWithLexicalScope} lexicalScope
+     */
+    preserveName (identifierName: string, lexicalScope: TNodeWithLexicalScope): void;
 }