+++ test/dev/dev.ts
@@ -7,15 +7,15 @@ import { NO_ADDITIONAL_NODES_PRESET } from '../../src/options/presets/NoCustomNo
 
     let obfuscatedCode: string = JavaScriptObfuscator.obfuscate(
         `
-            function foo({bar = ''}) {
-                return bar;
-            }
-            
-            console.log(foo({bar: 1}));
+            const sdasdas = '123';
         `,
         {
             ...NO_ADDITIONAL_NODES_PRESET,
-            compact: false
+            compact: false,
+            identifierNamesGenerator: 'dictionary',
+            identifiersDictionary: ['a', 'b', 'c', 'd', 'e'],
+            stringArray: true,
+            stringArrayThreshold: 1
         }
     ).getObfuscatedCode();